Refactor client to remove agent tools
